legacy_archive/requirements.txt
google-generativeai
structlog
python-dotenv
ddgs
feedparser
supabase
requests
streamlit-agraph
